team-website: 레거시 fixed form 코드 전부 제거 💣
