+++ _packages/@karrotmarket/gatsby-theme-team-website/src/templates/JobApplicationPage.tsx
@@ -6,15 +6,21 @@ import { GatsbySeo } from 'gatsby-plugin-next-seo';
 import { rem } from 'polished';
 import { required } from '@cometjs/core';
 import type { PropOf, RefOf } from '@cometjs/react-utils';
+import { mapAbstractType } from '@cometjs/graphql-utils';
 
 import _PageTitle from '../components/PageTitle';
 import _FormField from '../components/FormField';
+import FileAttachmentField from '../components/formField/FileAttachmentField';
+import ShortTextField from '../components/formField/ShortTextField';
+import LongTextField from '../components/formField/LongTextField';
+import SingleSelectField from '../components/formField/SingleSelectField';
+import MultiSelectField from '../components/formField/MultiSelectField';
+import YesNoField from '../components/formField/YesNoField';
+import TermsField from '../components/formField/TermsField';
 import Button from '../components/Button';
 import _Spinner from '../components/Spinner';
 
-import type { ApplicationForm } from '../utils/applicationForm';
-import { makeClient, makeEndpoint } from '../utils/applicationForm';
-import * as Base64 from '../utils/base64';
+import { makeNewEndpoint } from '../utils/applicationForm';
 
 import messages from './jobApplicationPage/messages';
 
@@ -28,14 +34,39 @@ export const query = graphql`
   ) {
     ...TeamWebsite_DefaultLayout_query
     ...TeamWebsite_JobPostLayout_query
-
     jobPost(id: { eq: $id }) {
       ghId
       title
       boardToken
       portfolioRequired
+      parentJob {
+        questions {
+          __typename
+          name
+          label
+          required
+          description
+          ...on GreenhouseJobBoardJobQuestionForYesNo {
+            options {
+              label
+              value
+            }
+          }
+          ...on GreenhouseJobBoardJobQuestionForSingleSelect {
+            options {
+              label
+              value
+            }
+          }
+          ...on GreenhouseJobBoardJobQuestionForMultiSelect {
+            options {
+              label
+              value
+            }
+          }
+        }
+      }
     }
-
     privacyPolicy: prismicTermsAndConditions(
       uid: { eq: "job-application-privacy" }
       lang: { eq: $locale }
@@ -47,7 +78,6 @@ export const query = graphql`
         }
       }
     }
-
     sensitiveInfoPolicy: prismicTermsAndConditions(
       uid: { eq: "job-application-sensitive" }
       lang: { eq: $locale }
@@ -117,12 +147,6 @@ const FormField = styled(_FormField, {
   marginBottom: rem(32),
 });
 
-const FormHelpText = styled('p', {
-  color: '$gray600',
-  fontSize: '$caption1',
-  marginBottom: rem(48),
-});
-
 const Spinner = styled(_Spinner, {
   height: '50%',
 });
@@ -135,6 +159,11 @@ const greenhouseAcceptedMimeTypes = [
   'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
 ];
 
+const makeEndpoint = (boardToken: string, jobId: string): string => {
+  const host = process.env.GATSBY_JOB_APPLICATION_FORM_HOST || 'http://localhost:8787';
+  return `${host.replace(/\/$/, '')}/boards/${boardToken}/jobs/${jobId}/application/proxy`;
+};
+
 const JobApplicationPage: React.FC<JobApplicationPageProps> = ({
   data,
 }) => {
@@ -142,15 +171,13 @@ const JobApplicationPage: React.FC<JobApplicationPageProps> = ({
 
   const [state, dispatch] = React.useReducer(reducer, initialState);
 
-  const jobApplicationFormEndpoint = makeEndpoint(
-    process.env.GATSBY_JOB_APPLICATION_FORM_HOST || 'http://localhost:8787',
-    data.jobPost.boardToken,
-    data.jobPost.ghId,
-  );
+  const jobApplicationFormEndpoint = makeEndpoint(data.jobPost.boardToken, data.jobPost.ghId);
 
   type FormRef = RefOf<typeof Form>;
   const formRef = React.useRef<FormRef>(null);
 
+  // Note: Progressive Enhancement
+  // 사실 이거 없어도 기본 폼으로 100% 동작함
   type SubmitHandler = NonNullable<PropOf<typeof Form, 'onSubmit'>>;
   const handleSubmit: SubmitHandler = e => {
     e.preventDefault();
@@ -160,50 +187,27 @@ const JobApplicationPage: React.FC<JobApplicationPageProps> = ({
     }
 
     const formData = new FormData(formRef.current);
-
-    const applicationForm: ApplicationForm = {
-      phoneNumber: formData.get('phone_number') as string,
-      email: formData.get('email') as string,
-      name: formData.get('name') as string,
-      resume: formData.get('resume') as File,
-      portfolio: formData.get('portfolio') as File,
-      veterans: formData.get('veterans') as string,
-      disability: formData.get('disability') as string,
-      alternativeCivilian: formData.get('alternative_civilian') as string,
-    };
-
+    
     (async () => {
       required(data.jobPost);
 
-      const client = makeClient({
-        fetch,
-        endpoint: jobApplicationFormEndpoint,
-        encodeFile: async file => {
-          const content = await Base64.fromBlob(file);
-          return { content, filename: file.name };
-        },
-        portfolioRequired: data.jobPost.portfolioRequired,
-      });
-
-      if (client.validate(applicationForm)) {
-        dispatch('FETCH_START');
-
-        try {
-          const response = await client.submit(applicationForm);
-
-          if (response.ok) {
-            dispatch('FETCH_COMPLETE');
-            window.alert(messages.alert_completed);
-          } else {
-            dispatch('INVALID');
-            const message = await response.text();
-            window.alert(message);
-          }
-        } catch (e) {
-          console.error(e);
-          window.alert(messages.alert_failed);
+      dispatch('FETCH_START');
+      try {
+        const response = await fetch(jobApplicationFormEndpoint, {
+          method: 'POST',
+          body: formData,
+        });
+        if (response.ok) {
+          dispatch('FETCH_COMPLETE');
+          window.alert(messages.alert_completed);
+        } else {
+          dispatch('INVALID');
+          const message = await response.text();
+          window.alert(message);
         }
-      } else {
+      } catch (e) {
+        console.error(e);
+        window.alert(messages.alert_failed);
         dispatch('INVALID');
       }
     })();
@@ -217,121 +221,137 @@ const JobApplicationPage: React.FC<JobApplicationPageProps> = ({
     }
   }, [state]);
 
+  const portfolioField = data.jobPost.parentJob.questions.find(question => question.name === 'cover_letter');
+
   return (
     <Form
       ref={formRef}
       method="post"
+      encType="multipart/form-data"
       action={jobApplicationFormEndpoint}
       onSubmit={handleSubmit}
     >
       <GatsbySeo noindex />
       <FormField
-        variants={{ type: 'text' }}
-        name="name"
+        as={ShortTextField}
+        name="first_name"
         label={messages.field_name_label}
         placeholder={messages.field_name_placeholder}
         required
       />
+      {/* Treat the first_name as fullname */}
+      <input type="hidden" name="last_name" value={"\u200b"} />
       <FormField
-        variants={{ type: 'tel' }}
-        name="phone_number"
+        as={ShortTextField}
+        type="tel"
+        name="phone"
         label={messages.field_phone_label}
         placeholder={messages.field_phone_placeholder}
         required
       />
       <FormField
-        variants={{ type: 'email' }}
+        as={ShortTextField}
+        type="email"
         name="email"
         label={messages.field_email_label}
         placeholder={messages.field_email_placeholder}
         required
       />
       <FormField
-        variants={{
-          type: 'file',
-          accepts: greenhouseAcceptedMimeTypes,
-        }}
+        as={FileAttachmentField}
         name="resume"
+        accepts={greenhouseAcceptedMimeTypes}
         label={messages.field_resume_label}
         description={messages.field_resume_description}
         placeholder={messages.field_resume_placeholder}
         required
       />
-      <FormField
-        variants={{
-          type: 'file',
-          accepts: greenhouseAcceptedMimeTypes,
-        }}
-        name="portfolio"
-        label={messages.field_portfolio_label}
-        placeholder={messages.field_portfolio_placeholder}
-        description={messages.field_portfolio_description}
-        required={data.jobPost.portfolioRequired}
-      />
-      <FormField
-        variants={{
-          type: 'radio',
-          options: [
-            { label: '해당', value: 'skilled_industrial_personnel' },
-            { label: '비해당', value: 'no' },
-          ],
-          defaultValue: 'no',
-        }}
-        name="alternative_civilian"
-        label="산업기능요원"
-        required
-      />
-      <FormField
-        variants={{
-          type: 'select',
-          options: [
-            { label: '해당 없음', value: 'no' },
-            { label: '일반', value: 'normal' },
-            { label: '산재', value: 'industry' },
-            { label: '보훈', value: 'military' },
-          ],
-          defaultValue: 'no',
-        }}
-        label="장애사항"
-        name="disability"
-        required
-      />
-      <FormField
-        variants={{
-          type: 'radio',
-          options: [
-            { label: '대상', value: 'yes' },
-            { label: '비대상', value: 'no' },
-          ],
-          defaultValue: 'no',
-        }}
-        name="veterans"
-        label="보훈대상 여부"
-        required
-      />
-      <FormHelpText>
-        * 보훈 및 장애 사항은 채용 과정에서 불이익이 없습니다.
-      </FormHelpText>
+      {portfolioField && (
+        <FormField
+          as={FileAttachmentField}
+          accepts={greenhouseAcceptedMimeTypes}
+          name={portfolioField.name}
+          label={messages.field_portfolio_label}
+          description={messages.field_portfolio_description}
+          placeholder={messages.field_portfolio_placeholder}
+          required={portfolioField.required}
+        />
+      )}
+      {data.jobPost.parentJob.questions
+        // Note: Custom Question 만 따로 렌더링
+        .filter(question => question.name.startsWith('question'))
+        .map(question => mapAbstractType(question, {
+        GreenhouseJobBoardJobQuestionForShortText: question => (
+          <FormField
+            as={ShortTextField}
+            key={question.name}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+          />
+        ),
+        GreenhouseJobBoardJobQuestionForLongText: question => (
+          <FormField
+            as={LongTextField}
+            key={question.name}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+          />
+        ),
+        GreenhouseJobBoardJobQuestionForAttachment: question => (
+          <FormField
+            as={FileAttachmentField}
+            key={question.name}
+            accepts={greenhouseAcceptedMimeTypes}
+            placeholder={messages.custom_field_file_placeholder}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+          />
+        ),
+        GreenhouseJobBoardJobQuestionForYesNo: question => (
+          <FormField
+            as={YesNoField}
+            key={question.name}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+          />
+        ),
+        GreenhouseJobBoardJobQuestionForSingleSelect: question => (
+          <FormField
+            as={SingleSelectField}
+            key={question.name}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+            options={[...question.options]}
+          />
+        ),
+        GreenhouseJobBoardJobQuestionForMultiSelect: question => (
+          <FormField
+            as={MultiSelectField}
+            key={question.name}
+            name={question.name}
+            label={question.label}
+            required={question.required}
+            options={[...question.options]}
+          />
+        ),
+      }))}
       {data.privacyPolicy?.data?.content?.html && (
         <FormField
-          variants={{
-            type: 'terms',
-            terms: data.privacyPolicy.data.content.html,
-          }}
-          name="privacy"
+          as={TermsField}
+          terms={data.privacyPolicy.data.content.html}
           label={messages.terms_privacy_info}
-          required
         />
       )}
       {data.sensitiveInfoPolicy?.data?.content?.html && (
         <FormField
-          variants={{
-            type: 'terms',
-            terms: data.sensitiveInfoPolicy.data.content.html,
-          }}
-          name="sensitive"
+          as={TermsField}
+          terms={data.sensitiveInfoPolicy.data.content.html}
           label={messages.terms_sensitive_info}
-          required
         />
       )}
       <Button
